### 2024-08-01 #7064 @editorialbot
[PRE REVIEW]: ASTE: An artificial solver testing environment for partitioned coupling with preCICE
**Submitting author:** @davidscn (<a href="http://orcid.org/0000-0002-3487-9688">David Schneider</a>)
**Repository:** https://github.com/precice/aste
**Branch with paper.md** (empty if default branch): paper
**Version:** v3.2.0
