+++ tools/mapping-tester/generate.py
@@ -259,20 +259,21 @@ def createRunScript(outdir, path, case):
     )
 
 
-def setupCases(outdir, template, cases, exit):
+def setupCases(outdir, template, cases, exit, repetitions):
     casemap = {}
-    for case in cases:
-        folders = getCaseFolders(case)
-        casemap.setdefault(folders[0], []).append(folders[1:])
-        name = [outdir] + folders
-        path = os.path.join(*name)
-        config = os.path.join(path, "precice-config.xml")
+    for case in cases :
+        for rep in range(repetitions):
+            folders = getCaseFolders(case)+[str(rep)]
+            casemap.setdefault(folders[0], []).append(folders[1:])
+            name = [outdir] + folders
+            path = os.path.join(*name)
+            config = os.path.join(path, "precice-config.xml")
 
-        print(f"Generating {path}")
-        os.makedirs(path, exist_ok=True)
-        with open(config, "w") as config:
-            config.write(generateConfig(template, case))
-        createRunScript(outdir, path, case)
+            print(f"Generating {path}")
+            os.makedirs(path, exist_ok=True)
+            with open(config, "w") as config:
+                config.write(generateConfig(template, case))
+            createRunScript(outdir, path, case)
     print(f"Generated {len(cases)} cases")
 
     print(f"Generating master scripts")
@@ -323,7 +324,8 @@ def main(argv):
     if os.path.isdir(outdir):
         print('Warning: outdir "{}" already exisits.'.format(outdir))
 
-    setupCases(outdir, template, cases, args.exit)
+    repetitions = setup["general"].get("repetitions", 1)
+    setupCases(outdir, template, cases, args.exit, repetitions)
 
     return 0
 
