### 2024-08-07 e1d4551
Fix missing dois where possible
