+++ docs/paper.bib
@@ -26,16 +26,16 @@ publisher = {DaRUS},
 title = {{preCICE Distribution Version v2404.0}},
 year = {2024},
 version = {V1},
-doi = {10.18419/darus-4167},
-url = {https://doi.org/10.18419/darus-4167}
+doi = {10.18419/darus-4167}
 }
 
 @inproceedings{Schneider2023,
-author = {Schneider, D. and Schrader, T. and Uekermann, B.},
+author = { Schneider, D. and Schrader, T. and Uekermann, B.},
 title = {Data-parallel radial-basis-function interpolation in {preCICE}},
 year = {2023},
-Booktitle = {X International Conference on Computational Methods for Coupled Problems in Science and Engineering}}
-
+Booktitle = {X International Conference on Computational Methods for Coupled Problems in Science and Engineering},
+doi = {10.23967/c.coupled.2023.016}
+}
 
 @mastersthesis{Martin2022,
     author = {Martin, Boris Guillaume},
@@ -51,7 +51,7 @@ Booktitle = {X International Conference on Computational Methods for Coupled Pro
   type={Bachelor's thesis},
   institution = {University of Stuttgart},
   year={2022},
-  url = {http://dx.doi.org/10.18419/opus-12128}
+  doi = {10.18419/opus-12128}
 }
 
 @phdthesis{Lindner2019,
@@ -95,7 +95,8 @@ year="2020",
 publisher="Springer International Publishing",
 address="Cham",
 pages="271--300",
-isbn="978-3-030-47956-5"
+isbn="978-3-030-47956-5",
+doi="10.1007/978-3-030-47956-5_10"
 }
 
 @MISC{METIS,
