### 2024-08-07 #7064 @davidscn
I added now missing DOIs where possible. For the remaining entries, there are at most URLs given and no DOIs available. If that's a problem, I would consider to embed the mentioned URL into the plain text rather than making it a full citation.
