### 2024-08-08 0550f23
Enable Boost policy
