### 2024-08-08 7eddc3e
Fix linker errors when using boost from config
