### 2024-08-08 b9363f1
Add repetitions to mapping tester
