### 2024-08-08 0c4d7ee
Overhauled gatherstats
