+++ changelog-entries/194.md
@@ -1,0 +1,1 @@
+- Fixed linker errors when FindBoost uses the config.
