### 2024-08-08 fe77af0
Fix Boost CMake usage (#194)
* Enable Boost policy

* Fix linker errors when using boost from config

* Add changelog

* Request config

* Update 194.md
