+++ changelog-entries/194.md
@@ -1,0 +1,2 @@
+- Fixed linker errors when FindBoost uses the CMake Config.
+- Changed CMake to always find boost using its CMake Config.
