### 2024-08-08 d38af71
Fix SEGFAULT on missing mesh (#196)
* Fix SEGFAULT on missing mesh

* Add changelog
