### 2024-08-08 a10a30b
Add support for no output mesh
