### 2024-08-08 97c5b40
Add support for disabling writing mapped data
