+++ tools/mapping-tester/generate.py
@@ -100,7 +100,7 @@ def caseToSortable(case):
     return (kindCost, -mesha, -meshb)
 
 
-def createMasterRunScripts(casemap, dir, exit):
+def createMasterRunScripts(casemap, dir, exit, writeMapped):
     common = [
         "#!/bin/bash",
         "",
@@ -124,21 +124,22 @@ def createMasterRunScripts(casemap, dir, exit):
         [line + "\n" for line in content]
     )
 
-    # Generate master postprocessing script
-    if exit:
-        post = common + [
-            "${RUNNER} " + os.path.join(case, "postprocessall.sh || exit 1")
-            for case in casemap.keys()
-        ]
-    else:
-        post = common + [
-            "${RUNNER} " + os.path.join(case, "postprocessall.sh")
-            for case in casemap.keys()
-        ]
+    # Generate master postprocessing script for mapped meshes
+    if writeMapped:
+        if exit:
+            post = common + [
+                "${RUNNER} " + os.path.join(case, "postprocessall.sh || exit 1")
+                for case in casemap.keys()
+            ]
+        else:
+            post = common + [
+                "${RUNNER} " + os.path.join(case, "postprocessall.sh")
+                for case in casemap.keys()
+            ]
 
-    open(os.path.join(dir, "postprocessall.sh"), "w").writelines(
-        [line + "\n" for line in post]
-    )
+        open(os.path.join(dir, "postprocessall.sh"), "w").writelines(
+            [line + "\n" for line in post]
+        )
 
     for case, instances in casemap.items():
         # Generate master runner script
@@ -150,16 +151,18 @@ def createMasterRunScripts(casemap, dir, exit):
             [line + "\n" for line in content]
         )
 
-        # Generate master postprocessing script
-        post = common + [
-            "${RUNNER} " + os.path.join(*instance, "post.sh") for instance in instances
-        ]
-        open(os.path.join(dir, case, "postprocessall.sh"), "w").writelines(
-            [line + "\n" for line in post]
-        )
+        # Generate master postprocessing script for mapped meshes
+        if writeMapped:
+            post = common + [
+                "${RUNNER} " + os.path.join(*instance, "post.sh")
+                for instance in instances
+            ]
+            open(os.path.join(dir, case, "postprocessall.sh"), "w").writelines(
+                [line + "\n" for line in post]
+            )
 
 
-def createRunScript(outdir, path, case):
+def createRunScript(outdir, path, case, writeMapped):
     amesh = case["A"]["mesh"]["name"]
     aranks = case["A"]["ranks"]
     ameshLocation = os.path.relpath(
@@ -184,8 +187,8 @@ def createRunScript(outdir, path, case):
         os.path.join(outdir, "meshes", bmesh, str(branks), bmesh), path
     )
     mapped_data_name = case["function"] + "(mapped)"
-    bcmd = f'env {time_command} -f %M -a -o memory-B.log precice-aste-run -v -a -p B --data "{mapped_data_name}" --mesh {bmeshLocation} --output mapped || kill 0 &'
-
+    out_mesh = "--output mapped" if writeMapped else ""
+    bcmd = f'env {time_command} -f %M -a -o memory-B.log precice-aste-run -v -a -p B --data "{mapped_data_name}" --mesh {bmeshLocation} {out_mesh} || kill 0 &'
     if branks > 1:
         bcmd = "mpirun -n {} $ASTE_B_MPIARGS {}".format(branks, bcmd)
 
@@ -232,39 +235,40 @@ def createRunScript(outdir, path, case):
         [line + "\n" for line in wrapper]
     )
 
-    # Generate post processing script
-    post_content = [
[…]
