### 2024-08-08 1b3373e
Add support for disabling writing mapped data
