+++ tools/mapping-tester/generate.py
@@ -99,7 +99,7 @@ def caseToSortable(case):
     return (kindCost, -mesha, -meshb)
 
 
-def createMasterRunScripts(casemap, dir, exit):
+def createMasterRunScripts(casemap, dir, exit, writeMapped):
     common = [
         "#!/bin/bash",
         "",
@@ -123,21 +123,22 @@ def createMasterRunScripts(casemap, dir, exit):
         [line + "\n" for line in content]
     )
 
-    # Generate master postprocessing script
-    if exit:
-        post = common + [
-            "${RUNNER} " + os.path.join(case, "postprocessall.sh || exit 1")
-            for case in casemap.keys()
-        ]
-    else:
-        post = common + [
-            "${RUNNER} " + os.path.join(case, "postprocessall.sh")
-            for case in casemap.keys()
-        ]
+    # Generate master postprocessing script for mapped meshes
+    if writeMapped:
+        if exit:
+            post = common + [
+                "${RUNNER} " + os.path.join(case, "postprocessall.sh || exit 1")
+                for case in casemap.keys()
+            ]
+        else:
+            post = common + [
+                "${RUNNER} " + os.path.join(case, "postprocessall.sh")
+                for case in casemap.keys()
+            ]
 
-    open(os.path.join(dir, "postprocessall.sh"), "w").writelines(
-        [line + "\n" for line in post]
-    )
+        open(os.path.join(dir, "postprocessall.sh"), "w").writelines(
+            [line + "\n" for line in post]
+        )
 
     for case, instances in casemap.items():
         # Generate master runner script
@@ -149,16 +150,18 @@ def createMasterRunScripts(casemap, dir, exit):
             [line + "\n" for line in content]
         )
 
-        # Generate master postprocessing script
-        post = common + [
-            "${RUNNER} " + os.path.join(*instance, "post.sh") for instance in instances
-        ]
-        open(os.path.join(dir, case, "postprocessall.sh"), "w").writelines(
-            [line + "\n" for line in post]
-        )
+        # Generate master postprocessing script for mapped meshes
+        if writeMapped:
+            post = common + [
+                "${RUNNER} " + os.path.join(*instance, "post.sh")
+                for instance in instances
+            ]
+            open(os.path.join(dir, case, "postprocessall.sh"), "w").writelines(
+                [line + "\n" for line in post]
+            )
 
 
-def createRunScript(outdir, path, case):
+def createRunScript(outdir, path, case, writeMapped):
     amesh = case["A"]["mesh"]["name"]
     aranks = case["A"]["ranks"]
     ameshLocation = os.path.relpath(
@@ -178,8 +181,9 @@ def createRunScript(outdir, path, case):
         os.path.join(outdir, "meshes", bmesh, str(branks), bmesh), path
     )
     mapped_data_name = case["function"] + "(mapped)"
-    bcmd = 'env time -f %M -a -o memory-B.log precice-aste-run -v -a -p B --data "{}" --mesh {} --output mapped || kill 0 &'.format(
-        mapped_data_name, bmeshLocation
+    out_mesh = "--output mapped" if writeMapped else ""
+    bcmd = 'env time -f %M -a -o memory-B.log precice-aste-run -v -a -p B --data "{}" --mesh {} {} || kill 0 &'.format(
+        mapped_data_name, bmeshLocation, out_mesh
     )
     if branks > 1:
         bcmd = "mpirun -n {} $ASTE_B_MPIARGS {}".format(branks, bcmd)
@@ -227,39 +231,40 @@ def createRunScript(outdir, path, case):
         [line + "\n" for line in wrapper]
     )
 
-    # Generate post processing script
[…]
