### 2024-08-08 5c0c8fd
Add test for runtime profiling
