+++ examples/mapping_tester_runtime/clean.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+rm -f test-statistics.csv
+rm -fr ./case/
