### 2024-08-14 #7064 @asalmgren
Realistically I couldn't get to it until October.

On Fri, Aug 9, 2024 at 9:12 AM Patrick Diehl ***@***.***>
wrote:

> Hi @asalmgren <https://github.com/asalmgren> do you have time to review
> this paper?
>
