### 2024-08-16 #7064 @davidscn
> @davidscn could you please recommend some potential reviewers?

Not entirely sure about the prerequisites to be a reviewer. Of course, we could suggest again the reviewers from https://github.com/openjournals/joss-reviews/issues/5842, as they are somewhat familiar with the preCICE-verse and JOSS. Other than that I could maybe suggest `@hverhelst` or `@antRthorn`. Would that comply with the JOSS formalities?
