### 2024-08-29 9ebb14c
Clean the profiling directories instead of raw json files (preCICE v3 related)
