+++ examples/lci_2d/clean.sh
@@ -2,8 +2,7 @@
 set -e -x
 
 rm -f -r precice-run
-rm -f precice-*-events*.json
+rm -f -r precice-profiling
 rm -f *.log
 rm -f map*.vtk
 rm -f fine_mesh_*.vtk
-rm -f precice-*.json
