### 2024-08-30 483dabc
Add basic support for indirect mesh access
