### 2024-08-30 e19a48e
Add basic support for indirect mesh access
