+++ src/modes.cpp
@@ -146,6 +146,12 @@ void aste::runMapperMode(const aste::ExecutionContext &context, const OptionMap 
   const bool        isVector        = options["vector"].as<bool>();
   const std::string preciceConfig   = options["precice-config"].as<std::string>();
 
+  constexpr int conventionalReadData = -1;
+  // potentially make use of a indirect mapping using a prescribed batch size
+  const int batchSize = options.count("indirect-read")
+                            ? options["indirect-read"].as<int>()
+                            : conventionalReadData;
+
   addLogIdentity(participantName, context.rank);
   ASTE_INFO << "ASTE Running in mapping test mode";
 
@@ -186,7 +192,11 @@ void aste::runMapperMode(const aste::ExecutionContext &context, const OptionMap 
     asteConfiguration.participantName       = "B";
     asteConfiguration.preciceConfigFilename = preciceConfig;
     aste::asteInterface asteInterface;
-    asteInterface.meshName       = "B-Mesh";
+    if (batchSize == conventionalReadData) {
+      asteInterface.meshName = "B-Mesh";
+    } else {
+      asteInterface.meshName = "A-Mesh";
+    }
     asteInterface.meshFilePrefix = meshname;
     asteInterface.meshes         = aste::BaseName(meshname).findAll(context);
 
@@ -212,7 +222,14 @@ void aste::runMapperMode(const aste::ExecutionContext &context, const OptionMap 
   asteInterface.meshes.front().loadMesh(asteInterface.mesh, dim, requireConnectivity);
   asteInterface.meshes.front().loadData(asteInterface.mesh);
   ASTE_INFO << "The loaded mesh " << asteInterface.meshes.front().filename() << " contains: " << asteInterface.mesh.summary();
-  auto vertexIDs = aste::setupMesh(preciceInterface, asteInterface.mesh, asteInterface.meshName);
+
+  std::vector<int>    vertexIDs;
+  std::vector<double> coordinates;
+  if (batchSize == conventionalReadData) {
+    vertexIDs = aste::setupMesh(preciceInterface, asteInterface.mesh, asteInterface.meshName);
+  } else {
+    coordinates = aste::setupDirectMeshAccess(preciceInterface, asteInterface.mesh, asteInterface.meshName, context);
+  }
   ASTE_DEBUG << "Mesh setup completed on Rank " << context.rank;
 
   if (preciceInterface.requiresInitialData()) {
@@ -263,8 +280,13 @@ void aste::runMapperMode(const aste::ExecutionContext &context, const OptionMap 
     for (auto &asteInterface : asteConfiguration.asteInterfaces) {
       for (auto &meshdata : asteInterface.mesh.meshdata) {
         if (meshdata.type == aste::datatype::READ) {
-          meshdata.dataVector.resize(vertexIDs.size() * meshdata.numcomp);
-          preciceInterface.readData(asteInterface.meshName, "Data", vertexIDs, dt, meshdata.dataVector);
+          if (batchSize == conventionalReadData) {
+            meshdata.dataVector.resize(vertexIDs.size() * meshdata.numcomp);
+            preciceInterface.readData(asteInterface.meshName, "Data", vertexIDs, dt, meshdata.dataVector);
+          } else {
+            meshdata.dataVector.resize((coordinates.size() / preciceInterface.getMeshDimensions(meshname)) * meshdata.numcomp);
+            preciceInterface.mapAndreadData(asteInterface.meshName, "Data", coordinates, dt, meshdata.dataVector);
+          }
           ASTE_DEBUG << "Data read: " << asteInterface.mesh.previewData(meshdata);
         }
       }
