### 2024-08-30 bdb2a65
Add a test case for indirect access
