### 2024-08-30 ce28cdd
Add a test case for indirect access
