+++ examples/nn_indirect_access/clean.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+set -e -x
+
+rm -f -r precice-run
+rm -f -r precice-profiling
+rm -f *.log
+rm -f map*.vtk
+rm -f fine_mesh_*.vtk
