### 2024-08-30 302cf1e
Delete (post-processed) profiling and trace files in clean.sh
