+++ examples/nn/clean.sh
@@ -3,6 +3,8 @@ set -e -x
 
 rm -f -r precice-run
 rm -f -r precice-profiling
+rm -f profiling.json
+rm -f trace.json
 rm -f *.log
 rm -f map*.vtk
 rm -f fine_mesh_*.vtk
