### 2024-08-30 62a8755
Implement a batch-wise processing of data
