### 2024-08-30 573bbb3
Implement a batch-wise processing of data
