### 2024-08-30 c494633
Special treatment for batch size of zero
