### 2024-08-30 85b2c61
Special treatment for batch size of zero
