### 2024-08-30 #7127 @asalmgren
I won’t be able to start this until late September or early October.

A

Ann Almgren
Senior Scientist; Dept. Head, Applied Mathematics
Pronouns: she/her/hers

On Fri, Aug 30, 2024 at 2:15 AM Patrick Diehl ***@***.***>
wrote:

> Hi @asalmgren <https://github.com/asalmgren>, @hverhelst
> <https://github.com/hverhelst> how is your review going?
>
