### 2024-09-02 005aed0
Take batch size for mapping tester into account
