### 2024-09-02 0db004d
Take batch size for mapping tester into account
