### 2024-09-04 precice/aste#199 @hverhelst
Running examples
Hi all!

I've successfully built ASTE and I am trying to run some examples. However, it is not clear to me how to actually run them (or there are still some installation issues).
- From `./aste/examples` running `run-all.sh` gives me the following error (and more lines below)
```
- Running all examples cases...
- There are 7 case(s) found...
- Running test case replay_mode...
./run.sh: line 7: precice-aste-run: command not found
./run.sh: line 10: precice-aste-run: command not found
...
```
I do have the executable `precice-aste-run` and the others in my `build`  folder.
- From `./aste/` running `./examples/run-all.sh` tells me that there are 0 examples to run.

Could someone help me out?

Cheers,

Hugo
