### 2024-09-04 #7127 @hverhelst
I had some questions and published an issue [here](https://github.com/precice/aste/issues/199)
