### 2024-09-04 precice/aste#199 @davidscn
Certainly!

as you noted, `precice-aste-run` is an executable generated as part of compiling ASTE. By default, the executable is not discoverable by your OS. To make it discoverable, you have two options:

1) run `make install`: this will `install` all generated executables in what is specified in `CMAKE_INSTALL_PREFIX` (default is `usr/local`, which installes it system-wide and requires sudo permission, you don't have to take care of anything else). You can change the install path using CMake: example `cmake -D CMAKE_INSTALL_PREFIX="my/path"`. If you go for a custom location, make sure that the custom location is discoverable as well (see 2).
2) make (any) location discoverable by the OS: it depends a bit on the OS, but for Ubuntu, the following environment variables can be used to make the executable discoverable, e.g., in the `build` directory:

```bash
export PATH="${PATH}:my/sys/path/aste/build/"
export LD_LIBRARY_PATH="${LD_LIBRARY_PATH}:my/sys/path/aste/build/"
```

or in a custom installation directory (through `CMAKE_INSTALL_PREFIX`):

```bash
export PATH="${PATH}:my/prefix/bin"
export LD_LIBRARY_PATH="${LD_LIBRARY_PATH}:my/prefix/lib"
```

This reminds me that some of the examples require the tools in the mapping-tester, which we don't install by default, but we should:

```bash
export PATH="my/path/aste/tools/mapping-tester/:${PATH}"
```
would work as well, but this was still on my TODO list. I will open a fix.

Does this help?
