### 2024-09-04 precice/aste#199 @davidscn
> I do have the executable precice-aste-run and the others in my build folder.
    From ./aste/ running ./examples/run-all.sh tells me that there are 0 examples to run.

This one is interesting: The `run-all.sh` is mostly a "find and execute" script for all other `run.sh` scripts. When executed from the root directory of the repository, it fails to locate the `run.sh` scripts. I will fix that as well, thanks for pointing this out.
