### 2024-09-04 542f774
Make run-all script executable from anywhere
