### 2024-09-04 22b8a6c
Early exit in case the executables are not found
