### 2024-09-05 569dd2f
Hint to installation page
