### 2024-09-19 84020cf
Add macOS CI
