### 2024-09-19 f7ac907
Remove components not requires or conflicting (hdf5-mpi vs hdf5)
