+++ .github/workflows/aste_ci_mac.yml
@@ -39,9 +39,15 @@ jobs:
           sudo ninja install
           cd ../..
       - uses: actions/checkout@v4
+      - name: Set up Python virtual environment
+        run: |
+          python3 -m venv venv
+          echo "VIRTUAL_ENV=$(pwd)/venv" >> $GITHUB_ENV
+          echo "$VIRTUAL_ENV/bin" >> $GITHUB_PATH
       - name: Install example dependencies
         run: |
-          python3 -m pip install -r requirements.txt
+          source $VIRTUAL_ENV/bin/activate
+          python -m pip install -r requirements.txt
       - name: Prepare build directory
         run: mkdir build
       - name: Build project with clang
@@ -49,10 +55,13 @@ jobs:
         env:
           CC: clang
           CXX: clang++
+          PATH: ${{ env.VIRTUAL_ENV }}/bin:${{ env.PATH }}
         run: |
           cmake ..
           cmake --build .
       - name: Run tests
         working-directory: build
+        env:
+          PATH: ${{ env.VIRTUAL_ENV }}/bin:${{ env.PATH }}
         run: |
           ctest
