+++ .github/workflows/aste_ci_mac.yml
@@ -24,7 +24,6 @@ jobs:
         run: |
           brew update
           brew install cmake eigen boost openmpi pkg-config ninja vtk
-          python3 -m pip install --upgrade pip
       - name: Install preCICE
         run: |
           git clone https://github.com/precice/precice.git
@@ -34,16 +33,25 @@ jobs:
                 -DCMAKE_BUILD_TYPE=RelWithDebInfo \
                 -DPRECICE_FEATURE_MPI_COMMUNICATION=ON \
                 -DPRECICE_FEATURE_PETSC_MAPPING=OFF \
+                -DPRECICE_FEATURE_PYTHON_ACTIONS=OFF \
                 -G Ninja ..
           ninja
           sudo ninja install
           cd ../..
       - uses: actions/checkout@v4
+      - name: Setup python
+        uses: actions/setup-python@v5
+        with:
+         python-version: '3.12'
+         check-latest: true
+
       - name: Set up environment and install dependencies
         run: |
           python3 -m venv venv
           source venv/bin/activate
+          python3 -m pip install --upgrade pip
           pip install -r requirements.txt
+          pip install numpy polars
           echo "VIRTUAL_ENV=$(pwd)/venv" >> $GITHUB_ENV
           echo "$VIRTUAL_ENV/bin" >> $GITHUB_PATH
 
