+++ .github/workflows/aste_ci_mac.yml
@@ -52,7 +52,7 @@ jobs:
           source venv/bin/activate
           python3 -m pip install --upgrade pip
           pip install -r requirements.txt
-          pip install numpy polars
+          pip install numpy polars vtk
           echo "VIRTUAL_ENV=$(pwd)/venv" >> $GITHUB_ENV
           echo "$VIRTUAL_ENV/bin" >> $GITHUB_PATH
 
