### 2024-09-19 9e042db
Use default network from preCICE
