### 2024-09-19 f7bb59c
Add macOS CI for ASTE (#202)
* Add macOS CI

* Remove components not requires or conflicting (hdf5-mpi vs hdf5)

* Use venv for python dependencies

* ......

* ..

* Use sys site

* ...

* install gnu-time as well

* Use default network from preCICE
