### 2024-10-20 #7127 @asalmgren
Regarding the community guidelines -- I do not see a CONTRIBUTING.md or similar file that gives guidance to anyone wanting to contribute to the software or report bugs.  I do not see this information in the README file either.
