### 2024-10-20 #7127 @asalmgren
Regarding the installation instructions -- the dependencies are clearly identified.  However, if I follow the "Building and Installation" instructions I get the following error message.  I'm assuming that I need to have built precice before I can build aste but that should be clearly stated in the "Building and installation" sections of the instructions.

CMake Error at CMakeLists.txt:24 (find_package):
  By not providing "Findprecice.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "precice", but
  CMake did not find one.

  Could not find a package configuration file provided by "precice"
  (requested version 3.0) with any of the following names:

    preciceConfig.cmake
    precice-config.cmake

  Add the installation prefix of "precice" to CMAKE_PREFIX_PATH or set
  "precice_DIR" to a directory containing one of the above files.  If
  "precice" provides a separate development package or SDK, be sure it has
  been installed.
