### 2024-10-21 precice/aste#203 @davidscn
Missing `CONTRIBUTING` guidelines
As reported in https://github.com/openjournals/joss-reviews/issues/7127#issuecomment-2425288966
