### 2024-10-21 ebb45a8
Add changelog entry
