+++ docs/CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# Contributing to ASTE
+
+Thank you for considering contributing to our project! We welcome contributions of all kinds.
