### 2024-10-21 37f17fd
Add contributing guidelines
