+++ changelog-entries/205.md
@@ -1,1 +1,2 @@
 - Added a code of conduct from Covenant (see `CODE_OF_CONDUCT.md`).
+- Added contributing guidelines (see `docs/CONTRIBUTING.md`).
