### 2024-10-24 c08095e
Apply suggestions from code review
Co-authored-by: Benjamin Uekermann <benjamin.uekermann@ipvs.uni-stuttgart.de>
