### 2024-10-24 afb7406
Add community guidelines (#205)
* Add a code_of_conduct based on covenant

* Add changelog entry

* minor WIP

* Add contributing guidelines

* update changelog

* Apply suggestions from code review

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@ipvs.uni-stuttgart.de>

---------

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@ipvs.uni-stuttgart.de>
