### 2024-10-24 #7127 @davidscn
@asalmgren thank you for your review comments. We addressed both points in precice/aste#203 and precice/aste#204 respectively.
