### 2024-10-25 aed87c4
Add scoping support to gathering events
