+++ tools/mapping-tester/gatherstats.py
@@ -49,15 +49,10 @@ def statsFromTimings(dir):
                         stats["globalTime"] = row[-1]
                     if row[0] == "initialize":
                         stats["initializeTime"] = row[-1]
-                    if row[0].startswith("initialize/map") and row[0].endswith(
-                        "computeMapping.FromA-MeshToB-Mesh"
-                    ):
-                        computeMappingName = row[0]
+                    parts = row[-1].split("/")
+                    if parts[0] == "initialize" and parts[-1].startswith("map") and parts[-1].endswith("computeMapping.FromA-MeshToB-Mesh"):
                         stats["computeMappingTime"] = row[-1]
-                    if row[0].startswith("advance/map") and row[0].endswith(
-                        "mapData.FromA-MeshToB-Mesh"
-                    ):
-                        mapDataName = row[0]
+                    if parts[0] == "advance" and parts[-1].startswith("map") and parts[-1].endswith("mapData.FromA-MeshToB-Mesh"):
                         stats["mapDataTime"] = row[-1]
         except BaseException:
             pass
