+++ changelog-entries/209.md
@@ -1,0 +1,1 @@
+- Added documentation to the mapping tester.
