### 2024-10-25 ab0a415
Improve performance and robustness of gathering stats
