### 2024-10-25 815f28d
Improve performance and robustness of gathering stats
