+++ changelog-entries/210.md
@@ -1,0 +1,1 @@
+- Changed `gatherstats.py` of the mapping tester to run in parallel and aggregate all available data. Running post-processing scripts before is now optional.
