### 2024-10-25 0419cea
Use ASTE ci-image
