### 2024-10-25 2f1e281
Ensure clang is installed
