+++ .github/workflows/aste_ci.yml
@@ -31,6 +31,7 @@ jobs:
       CXX: ${{ matrix.CXX }}
       CXX_FLAGS: "-Werror -Wall -Wextra -Wno-unused-parameter"
       CTEST_OUTPUT_ON_FAILURE: "Yes"
+      CMAKE_BUILD_PARALLEL_LEVEL: 4
     steps:
       - name: setup system
         run: |
