### 2024-10-25 d60267b
Add a compare script
