### 2024-10-25 e08b82b
Add reference statistics
