+++ changelog-entries/209.md
@@ -1,1 +1,0 @@
-- Added documentation to the mapping tester.
