+++ changelog-entries/196.md
@@ -1,1 +1,0 @@
-- Fixed SEGFAULT in `precice-aste-run` if no mesh is found for the given name.
