### 2024-11-06 #7127 @davidscn
@diehlpk thank you for your comments. I will address the first two points as soon as possible. May I ask you to elaborate more in detail on the last point? We have an enumeration of where ASTE is/was used in the scientific context, including corresponding citations in the last paragraph of the 'Statement of Need'. The tool, however, is not integrated into other software tools as such.
