### 2024-11-07 0efa458
Alternative typesetting for C++
