### 2024-11-07 892283d
Add configurable partitioning algorithm
