+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [v3.3.0] - 2024-11-XX
+## [v3.3.0] - 2024-11-07
 
 ### Added
 
