+++ changelog-entries/215.md
@@ -1,0 +1,1 @@
+- Changed the minimum required CMake version to 3.22.
