### 2024-11-07 1a5b05c
Add small paragraph for similar tools
