+++ docs/paper.bib
@@ -20,6 +20,18 @@ ISSN = {1999-4893},
 DOI = {10.3390/a14060166}
 }
 
+@Article{Valcke2022,
+AUTHOR = {Valcke, Sophie and Piacentini, Andrea and Jonville, Gabriel},
+TITLE = {Benchmarking Regridding Libraries Used in Earth System Modelling},
+JOURNAL = {Mathematical and Computational Applications},
+VOLUME = {27},
+YEAR = {2022},
+NUMBER = {2},
+ARTICLE-NUMBER = {31},
+ISSN = {2297-8747},
+DOI = {10.3390/mca27020031}
+}
+
 @data{preciceDistribution,
 author = {Chen, Jun and Chourdakis, Gerasimos and Desai, Ishaan and Homs-Pons, Carme and Rodenberg, Benjamin and Schneider, David and Simonis, Frédéric and Uekermann, Benjamin and Davis, Kyle and Jaust, Alexander and Kelm, Mathis and Kotarsky, Niklas and Kschidock, Helena and Mishra, Durganshu and Mühlhäußer, Markus and Schrader, Timo Pierre and Schulte, Miriam and Seitz, Valentin and Signorelli, Joseph and van Zwieten, Gertjan and Vinnitchenko, Niklas and Vladimirova, Tina and Willeke, Leonard and Zonta, Elia},
 publisher = {DaRUS},
