### 2024-11-07 #7127 @davidscn
@diehlpk I addressed all points mentioned in https://github.com/openjournals/joss-reviews/issues/7127#issuecomment-2460343570, thanks again. May I ask you to check again?

> @davidscn I was more thinking if there are other tools like ASTE or is this tool unique.

This is a bit of a difficult question. I added a smaller paragraph at the very end of the 'statement of need'. Finding these software environments as standalone, citable and accessible tools is difficult. I mentioned two concrete resources I am aware of.
