### 2024-11-07 #7127 @diehlpk
@davidscn Thanks for the clarification.  I would say adding these references is sufficient.
