### 2024-11-08 2ba978b
Add warning regarding missing partitioning
