### 2024-11-08 ecc6b23
Reword to ensure scripts are included
