### 2024-11-12 70f98f0
Paper wording changes (#217)
* Change wording

* Reword VTK file origin in intro

* Mention other sources for VTK files

* Reword to ensure scripts are included

* Revert
