### 2024-11-12 #7127 @davidscn
@diehlpk I checked all points. Here are the information

- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. Archive the release on Zenodo/figshare/etc and post the DOI here.
Version: [v3.3.0](https://github.com/precice/aste/releases/tag/v3.3.0)
DOI: [10.18419/darus-4562](https://doi.org/10.18419/darus-4562)
