### 2024-11-12 #7127 @diehlpk
@davidscn could you please update the title on darus to match the one of the paper?

Currently, the title is ASTE Version v3.3.0.
