### 2024-11-12 #7127 @davidscn
> Done! version is now v.3.3.0

The version set here has an additional dot after the v as opposed to the actual version  v3.3.0

> Currently, the title is ASTE Version v3.3.0.

Sorry, I solely focused on the authors and thought it would be ambiguous to use the same for the data. I contacted DaRUS and asked for fix. The fix is now published as a new (major) version of the data set. The DOI is

DOI: [10.18419/darus-4562](https://doi.org/10.18419/darus-4562)
