### 2024-11-12 #7127 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
