### 2024-11-13 fe81813
minor changes in paper.md
