+++ docs/paper.bib
@@ -35,7 +35,7 @@ DOI = {10.3390/mca27020031}
 @data{preciceDistribution,
 author = {Chen, Jun and Chourdakis, Gerasimos and Desai, Ishaan and Homs-Pons, Carme and Rodenberg, Benjamin and Schneider, David and Simonis, Frédéric and Uekermann, Benjamin and Davis, Kyle and Jaust, Alexander and Kelm, Mathis and Kotarsky, Niklas and Kschidock, Helena and Mishra, Durganshu and Mühlhäußer, Markus and Schrader, Timo Pierre and Schulte, Miriam and Seitz, Valentin and Signorelli, Joseph and van Zwieten, Gertjan and Vinnitchenko, Niklas and Vladimirova, Tina and Willeke, Leonard and Zonta, Elia},
 publisher = {DaRUS},
-title = {{preCICE Distribution Version v2404.0}},
+title = {{preCICE} Distribution Version v2404.0},
 year = {2024},
 version = {V1},
 doi = {10.18419/darus-4167}
@@ -113,7 +113,7 @@ doi="10.1007/978-3-030-47956-5_10"
 
 @MISC{METIS,
 author = {George Karypis and Vipin Kumar},
-title = {{MeTis: Unstructured Graph Partitioning and Sparse Matrix Ordering System, Version 4.0}},
+title = {{MeTis}: Unstructured Graph Partitioning and Sparse Matrix Ordering System, Version 4.0},
 howpublished = {\url{http://www.cs.umn.edu/~metis}},
 year = {2009},
 organization = {University of Minnesota, Minneapolis, MN},
@@ -121,7 +121,7 @@ organization = {University of Minnesota, Minneapolis, MN},
 
 @misc{lohmann2023,
   author       = {Niels Lohmann},
-  title        = {{JSON for Modern C++}},
+  title        = {{JSON} for Modern {C}++},
   year         = 2023,
   month        = nov,
   version      = {3.11.3},
