### 2024-11-13 #7127 @danielskatz
@davidscn - As track chair, I've suggested some changes in https://github.com/precice/aste/pull/218 - please merge this or let me know what you disagree with, then we can proceed.
