### 2024-11-13 96e87ab
changes in JOSS paper (#218)
* minor changes in paper.md

* Update paper.bib

* Rephrase entire sentence about multi-physics as suggestes

---------

Co-authored-by: David Schneider <david.schneider@ipvs.uni-stuttgart.de>
