### 2024-11-14 #7127 @danielskatz
Congratulations to @davidscn (David Schneider) and co-authors on your publication!!

JOSS depends on volunteers; we couldn't do this without you
